uart: update to last version

+++ include/uart.h
@@ -44,13 +44,13 @@
 /******************************************************************************/
 
 struct uart_init_par {
-	uint32_t	device_id;
+	uint8_t	device_id;
 	uint32_t 	baud_rate;
 	void 		*extra;
 };
 
 struct uart_desc {
-	uint32_t 	device_id;
+	uint8_t 	device_id;
 	uint32_t 	baud_rate;
 	void 		*extra;
 };
